requirements.txt
numpy~=1.23.5
Pillow~=9.2.0
nibabel~=3.2.2